+++ src/main/java/ii/cipriantarlev/marketmanagementapi/measuringunit/MeasuringUnitRepository.java
@@ -5,6 +5,6 @@ package ii.cipriantarlev.marketmanagementapi.measuringunit;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MeasuringUnitRepository extends JpaRepository<MeasuringUnit, Integer> {
+public interface MeasuringUnitRepository extends JpaRepository<MeasuringUnit, Long> {
 
 }
