+++ src/main/java/ii/cipriantarlev/marketmanagementapi/vat/VatRepository.java
@@ -5,6 +5,6 @@ package ii.cipriantarlev.marketmanagementapi.vat;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VatRepository extends JpaRepository<Vat, Integer> {
+public interface VatRepository extends JpaRepository<Vat, Long> {
 
 }
